Create printer.cfg
This is the printer.cfg file for my QIDI XMAX3. This is a custom FREEDI full klipper rework of the .cfg file. Ive added a BTT sensor, and a Shelly3 PRO. I added shell_command through KIAUH to change power states.  I added to the MOONRAKER.conf to enable the SHUTDOWN/REBOOT host commands.